requirements.txt
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
faiss-cpu==1.8.0
sentence-transformers==2.6.1
transformers==4.38.2
pymupdf==1.23.26        
rank_bm25==0.2.2
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
langchain-text-splitters==0.0.1
groq>=0.9.0
flask==3.0.3
easyocr==1.7.1
pillow==10.4.0
pymupdf==1.23.26
rank_bm25==0.2.2
flask==3.0.3

python-dotenv==1.0.1